requirements.txt
numpy
pandas
scikit-learn
matplotlib
pillow
joblib
seaborn
notebook
